TP2/requirements.txt

aiohttp
beautifulsoup4
lxml
Pillow
aiofiles
# Optional for real screenshots (choose one):
# selenium
# webdriver-manager
# playwright
